+++ src/domain/factories/journalism/article_tags/mod.rs
@@ -1,0 +1,4 @@
+pub mod create_article_tag_service_factory;
+pub mod delete_article_tag_service_factory;
+pub mod fetch_many_article_tags_service_factory;
+pub mod update_article_tag_service_factory;