+++ src/infra/http/controllers/sessions_controller.rs
@@ -6,7 +6,7 @@ use log::info;
 use serde_json::json;
 use validator::Validate;
 
-use crate::domain::factories::authenticate_user_service_factory;
+use crate::domain::factories::identity::authenticate_user_service_factory;
 use crate::domain::services::identity::authenticate_user_service::AuthenticateUserParams;
 use crate::error::DomainError;
 use crate::infra::extensions::validator::IntoDomainError;