+++ src/infra/http/controllers/team_users_controller.rs
@@ -1,10 +1,10 @@
 use super::controller::ControllerTrait;
 use super::AppResponse;
 use crate::core::pagination::DEFAULT_PER_PAGE;
-use crate::domain::factories::create_team_user_service_factory;
-use crate::domain::factories::delete_team_user_service_factory;
-use crate::domain::factories::fetch_many_team_users_service_factory;
-use crate::domain::factories::update_team_user_service_factory;
+use crate::domain::factories::teams::{
+    create_team_user_service_factory, delete_team_user_service_factory,
+    fetch_many_team_users_service_factory, update_team_user_service_factory,
+};
 use crate::domain::repositories::team_user_repository::TeamUserQueryType;
 use crate::domain::services::teams::{
     create_team_user_service::CreateTeamUserParams, delete_team_user_service::DeleteTeamUserParams,